refactor kinesis IT
