+++ integration-tests/src/test/java/org/apache/druid/tests/indexer/ITKafkaIndexingServiceTest.java
@@ -19,45 +19,7 @@
 
 package org.apache.druid.tests.indexer;
 
-import org.apache.druid.java.util.common.StringUtils;
-import org.apache.druid.java.util.common.logger.Logger;
-import org.apache.druid.testing.guice.DruidTestModuleFactory;
-import org.apache.druid.tests.TestNGGroup;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.DataProvider;
-import org.testng.annotations.Guice;
-import org.testng.annotations.Test;
-
-@Test(groups = TestNGGroup.KAFKA_INDEX)
-@Guice(moduleFactory = DruidTestModuleFactory.class)
-public class ITKafkaIndexingServiceTest extends AbstractKafkaIndexerTest
+abstract class ITKafkaIndexingServiceTest extends AbstractIndexerTest
 {
-  private static final Logger LOG = new Logger(ITKafkaIndexingServiceTest.class);
-  private static final String DATASOURCE = "kafka_indexing_service_test";
 
-  @DataProvider
-  public static Object[][] testParams()
-  {
-    return new Object[][]{
-        {"legacy_parser"},
-        {"input_format"}
-    };
-  }
-
-  @Test(dataProvider = "testParams")
-  public void testKafka(String param)
-  {
-    final String supervisorSpecPath = "legacy_parser".equals(param)
-                                      ? INDEXER_FILE_LEGACY_PARSER
-                                      : INDEXER_FILE_INPUT_FORMAT;
-    LOG.info("Starting test: ITKafkaIndexingServiceTest");
-    doKafkaIndexTest(StringUtils.format("%s_%s", DATASOURCE, param), supervisorSpecPath, false);
-  }
-
-  @AfterMethod
-  public void afterClass()
-  {
-    LOG.info("teardown");
-    doTearDown();
-  }
 }
